+++ packages/settings-library/src/settings_library/application.py
@@ -1,8 +1,7 @@
-from models_library.basic_types import BootModeEnum
 from pydantic import Field, PositiveInt
 
 from .base import BaseCustomSettings
-from .basic_types import BuildTargetEnum
+from .basic_types import BootModeEnum, BuildTargetEnum
 
 
 class BaseApplicationSettings(BaseCustomSettings):
