+++ packages/settings-library/src/settings_library/basic_types.py
@@ -1,52 +1,26 @@
-#
-# NOTE: This file copies some of the types from models_library.basic_types
-#       This is a minor evil to avoid the maintenance burden that creates
-#       an extra dependency to a larger models_library (intra-repo library)
-
-from enum import StrEnum
 from typing import Annotated, TypeAlias
 
+from common_library.basic_types import BootModeEnum, BuildTargetEnum, LogLevel
 from pydantic import Field, StringConstraints
+
+assert LogLevel  # nosec
+assert BootModeEnum  # nosec
+assert BuildTargetEnum  # nosec
+
+__all__: tuple[str, ...] = (
+    "LogLevel",
+    "BootModeEnum",
+    "BuildTargetEnum",
+)
+
 
 # port number range
 PortInt: TypeAlias = Annotated[int, Field(gt=0, lt=65535)]
+RegisteredPortInt: TypeAlias = Annotated[int, Field(gt=1024, lt=65535)]
 
 
 # e.g. 'v5'
 VersionTag: TypeAlias = Annotated[str, StringConstraints(pattern=r"^v\d$")]
-
-
-class LogLevel(StrEnum):
-    DEBUG = "DEBUG"
-    INFO = "INFO"
-    WARNING = "WARNING"
-    ERROR = "ERROR"
-
-
-class BootMode(StrEnum):
-    """
-    Values taken by SC_BOOT_MODE environment variable
-    set in Dockerfile and used during docker/boot.sh
-    """
-
-    DEFAULT = "default"
-    LOCAL = "local-development"
-    DEBUG = "debug"
-    PRODUCTION = "production"
-    DEVELOPMENT = "development"
-
-
-class BuildTargetEnum(StrEnum):
-    """
-    Values taken by SC_BUILD_TARGET environment variable
-    set in Dockerfile that defines the stage targeted in the
-    docker image build
-    """
-
-    BUILD = "build"
-    CACHE = "cache"
-    PRODUCTION = "production"
-    DEVELOPMENT = "development"
 
 
 # non-empty bounded string used as identifier
@@ -54,4 +28,3 @@ class BuildTargetEnum(StrEnum):
 IDStr: TypeAlias = Annotated[
     str, StringConstraints(strip_whitespace=True, min_length=1, max_length=50)
 ]
-RegisteredPortInt: TypeAlias = Annotated[int, Field(gt=1024, lt=65535)]
